+++ app/controllers/applicants_controller.rb
@@ -108,8 +108,6 @@ class ApplicantsController < ApplicationController
       applicant_scope = applicant_scope.where(enrolled: false)
     end
 
-    applicant_scope = applicant_scope.deprecated if params[:deprecated] == '1'
-
     if params[:format] == 'csv'
       if applicant_scope.count == 0
         redirect_to applicants_path, alert: 'No data was exported since no applicants matched the specified filters.'
@@ -202,8 +200,6 @@ class ApplicantsController < ApplicationController
       params[:applicant][:urm_types] ||= [] if params[:set_urm_types] == '1'
       params[:applicant][:laboratories] ||= [] if params[:set_laboratories] == '1'
       params[:applicant][:transition_position] ||= [] if params[:set_transition_position] == '1'
-
-      params[:applicant][:degree_types] ||= [] if params[:set_degree_types] == '1' # Will be removed when degree_types is removed
 
       if current_user and current_user.administrator?
         params[:applicant][:admin_update] = '1'